Committing changes to snowflake in Append Mode
In Append Mode the connection is writing to the table intended but not committing in the end (also autocommit is set to false) as a result, the results are not reflected in the table.

+++ src/main/scala/net/snowflake/spark/snowflake/io/StageWriter.scala
@@ -131,12 +131,16 @@ private[io] object StageWriter {
         }
         else
           conn.renameTable(table.name, tempTable.name)
+      } else {
+        conn.commit() 
       }
     } catch {
       case e: Exception =>
         // snowflake-todo: try to provide more error information,
         // possibly from actual SQL output
-        if (targetTable == tempTable) conn.dropTable(tempTable.name)
+        if (saveMode == SaveMode.Overwrite && params.useStagingTable) {
+          if (targetTable == tempTable) conn.dropTable(tempTable.name)
+        }
         log.error("Error occurred while loading files to Snowflake: " + e)
         throw e
     }
